Map Component Refactor
