+++ src/clickable-marker-map/index.js
@@ -1,0 +1,1 @@
+export { default, baseClickableMarkerMapPropTypes } from "./ClickableMarkerMap";
