+++ src/clickable-region-map/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ClickableRegionMap";
