+++ src/state-judicial-district-map/StateJudicialDistrictMap.js
@@ -3,7 +3,7 @@ import React from "react";
 import styled from "styled-components";
 import judicialDistrictsTopology from "../assets/maps/us_nd_judicial_districts.json";
 import { DEFAULT_TENANT, OTHER_LABEL, TENANTS, OTHER } from "../constants";
-import StateMap from "../state-map";
+import ClickableRegionMap from "../clickable-region-map";
 
 const ASPECT_RATIO = TENANTS[DEFAULT_TENANT].aspectRatio;
 
@@ -72,12 +72,12 @@ export default function StateJudicialDistrictMap({
 
   return (
     <StateJudicialDistrictMapWrapper>
-      <StateMap
+      <ClickableRegionMap
         aspectRatio={ASPECT_RATIO}
         LabelComponent={DistrictLabel}
         locationId={currentLocation}
         onRegionClick={onLocationClick}
-        stateTopology={judicialDistrictsTopology}
+        topology={judicialDistrictsTopology}
         width={width}
       />
       {otherRecord && (
